+++ Plugin/OptIn.php
@@ -108,22 +108,33 @@ class OptIn
             $customer = $this->customerSession->getCustomer();
         }
 
-        $payload = [
+        $store = $this->storeManager->getStore($subscriber->getStoreId());
+        $storeGroup = $store->getGroup();
+        $storeWebsite = $store->getWebsite();
+
+        $data = [
             'email' => $subscriber->getSubscriberEmail(),
             'customer_id' => $customer !== null ? $customer->getId() : "",
             'customer_group' => $customer !== null
                 ? $this->dataHelper->getCustomerGroupName((int) $customer->getGroupId())
                 : 'Guest',
-            'store' => $this->storeManager->getStore($subscriber->getStoreId())->getName(),
+            'store' => $store->getName(),
+            'store_group' => $storeGroup->getName(),
+            'store_website' => $storeWebsite->getName(),
             'first_name' => $customer !== null ? $customer->getFirstname() : '',
             'last_name' => $customer !== null ? $customer->getLastname() : '',
         ];
 
         // Fire opt-in request to Smaily.
+        $workflowId = $this->config->getSubscriberOptInWorkflowId($website);
+
+        $payload = ['addresses' => [$data]];
+        if ($workflowId > 0) {
+            $payload['autoresponder'] = $workflowId;
+        }
+
         $response = $this->dataHelper->getSmailyApiClient($website)
-            ->post('/api/autoresponder.php', [
-                'addresses' => [$payload],
-            ]);
+            ->post('/api/autoresponder.php', $payload);
 
         if ((int) $response['code'] !== 101) {
             $this->logger->error('Smaily opt-in API request failed', ['payload' => $payload, 'response' => $response]);